+++ api/pom.xml
@@ -27,12 +27,12 @@
         <version>1.0.2</version>
     </parent>
 
-    <groupId>jakarta.enterprise.concurrent</groupId>
-    <artifactId>jakarta.enterprise.concurrent-api</artifactId>
-    <version>1.2-SNAPSHOT</version>
-    <name>javax.enterprise.concurrent-api</name>
-    <description>JSR 236:Concurrency Utilities for Java EE</description>
-    <url>https://github.com/javaee/concurrency-ee-spec</url>
+    <groupId>jakarta.concurrency</groupId>
+    <artifactId>jakarta-concurrency-api</artifactId>
+    <version>2.0-SNAPSHOT</version>
+    <name></name>
+    <description>Jakarta Concurrency</description>
+    <url>https://github.com/eclipse-ee4j/concurrency-api</url>
 
      <licenses>
         <license>
@@ -63,19 +63,19 @@
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 
-        <api_package>javax.enterprise.concurrent</api_package>
+        <api_package>jakarta.concurrency</api_package>
         <non.final>false</non.final>
         <last.spec_version>0.0</last.spec_version>
-        <next.spec_version>1.2</next.spec_version>
+        <next.spec_version>2.0</next.spec_version>
         <spec.version>${next.spec_version}</spec.version>
         <new.spec.version />
         <build_number />
-        <packages.export>javax.enterprise.concurrent.*; version=${spec.bundle.version}</packages.export>
+        <packages.export>jakarta.concurrency.*; version=${spec.bundle.version}</packages.export>
     </properties>
 
     <organization>
-        <name>Oracle Corporation</name>
-        <url>http://www.oracle.com/</url>
+        <name>Eclipse Foundation</name>
+        <url>http://www.eclipse.org/</url>
     </organization>
 
     <issueManagement>
@@ -143,6 +143,7 @@
             </plugins>
         </pluginManagement>
         <plugins>
+            <!--
             <plugin>
                 <groupId>org.glassfish.build</groupId>
                 <artifactId>spec-version-maven-plugin</artifactId>
@@ -168,6 +169,7 @@
                     </execution>
                 </executions>
             </plugin>
+            -->
             <!--
               This plugin is reponsible for packaging artifacts
               as OSGi bundles.  Please refer to
@@ -239,33 +241,7 @@
                 <version>2.8</version>
                 <configuration>
 		    <additionalparam>-Xdoclint:none</additionalparam>
-                    <doctitle>Concurrency Utilities for Java EE ${project.version} API Specification</doctitle>
-                    <bottom>
-<![CDATA[Copyright &#169; 1996-2013,
-    <a href="http://www.oracle.com">Oracle</a>
-    and/or its affiliates. All Rights Reserved.
-    Use is subject to
-    <a href="{@docRoot}/doc-files/speclicense.html" target="_top">license terms</a>.
-]]>
-                    </bottom>
-                    <!--javaApiLinks>
-                        <property>
-                            <name>api_1.3</name>
-                            <value>http://download.oracle.com/javase/1.3/docs/api/</value>
-                        </property>
-                        <property>
-                            <name>api_1.4</name>
-                            <value>http://download.oracle.com/javase/1.4.2/docs/api/</value>
-                        </property>
-                        <property>
-                            <name>api_1.5</name>
-                            <value>http://download.oracle.com/javase/1.5.0/docs/api/</value>
-                        </property>
-                        <property>
-                            <name>api_1.6</name>
-                            <value>http://download.oracle.com/javase/6/docs/api/</value>
-                        </property>
-                    </javaApiLinks-->
+                    <doctitle>Jakarta Concurrency ${project.version} API Specification</doctitle>
                 </configuration>
                 <executions>
                     <execution>
