+++ api/src/main/java/jakarta/concurrency/ManagedScheduledExecutorService.java
@@ -14,7 +14,7 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
 
-package javax.enterprise.concurrent;
+package jakarta.concurrency;
 
 import java.util.concurrent.ScheduledExecutorService;
