+++ api/src/main/java/jakarta/concurrency/Trigger.java
@@ -14,7 +14,7 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
 
-package javax.enterprise.concurrent;
+package jakarta.concurrency;
 
 /**
  * Triggers allow application developers to plug in rules for when
